즐겨찾기 등록 아이콘 수정

+++ client/src/components/Ingredients/TableItem/index.tsx
@@ -5,24 +5,23 @@ import useTextAreaItem from './useTableItem';
 import * as S from './style';
 import { deleteFavorites } from '@contexts/user/actions';
 import favoriteAPI from '@lib/apis/favorite';
-export interface DataType {
-  id: number;
-  title: string;
-  latex: string;
-}
+import { FavoriteItem } from '@contexts/user/types';
 
 interface TableItemType {
   headerTitle: string;
   headerLatex: string;
-  data: DataType[];
+  data: FavoriteItem[];
 }
 
 function TableItem({ headerTitle, headerLatex, data }: TableItemType) {
   const { onChangeHandler } = useTextAreaItem();
   const dispatch = useDispatch();
-  const removeHandler = async (id: number) => {
-    dispatch(deleteFavorites(id));
-    await favoriteAPI.deleteFavorite(id);
+
+  const removeHandler = async (id?: number) => {
+    if (id) {
+      dispatch(deleteFavorites(id));
+      await favoriteAPI.deleteFavorite(id);
+    }
   };
   const dataList = data.map((item, index) => (
     <Table.Row key={index}>
